+++ docs/src/release_notes.md
@@ -5,6 +5,8 @@
 ### Breaking Changes in `NonlinearSolve.jl` v4
 
   - `ApproximateJacobianSolveAlgorithm` has been renamed to `QuasiNewtonAlgorithm`.
+  - Preconditioners for the linear solver needs to be specified with the linear solver
+    instead of `precs` keyword argument.
   - See [common breaking changes](@ref common-breaking-changes-v4v2) below.
 
 ### Breaking Changes in `SimpleNonlinearSolve.jl` v2